// src/Missions/constants/enums.ts
export enum MissionItems {
    CAR_PARTS = 'car-parts',
    CORPSE = 'corpse',
    ELECTRICAL_PARTS = 'electrical-parts',
    FOOD = 'food',
    MEDICAL = 'medical',
    MUNITIONS = 'munitions',
}

export enum ReportLocations {
    CEMETARY = 'cemetery',
    HOSPITAL = 'hospital',
    MILITARY_BASE = 'military-base',
    MYSTERIOUS_BUILDING = 'mysterious-building',
    POWER_STATION = 'power-station',
    RADIO_HQ = 'radio-hq',
    WAREHOUSE = 'warehouse',
}

export enum MissionTypes {
    PICKUP_DROPOFF = 'pickup-dropoff',
    REPORT = 'report',
}

export enum MissionActions {
    PICKUP = 'pick-up',
    DROPOFF = 'drop-off',
    REPORT = 'report',
}